mda.cpp, eis_hgb107x: Ericsson High resolution Monochrome Graphics Board 1070 split out in its own file as being an oddity
